requirements.txt
anthropic>=0.39.0
manim>=0.18.0
moviepy>=1.0.3
python-dotenv>=1.0.0
lmnt>=1.1.3
httpx>=0.28.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
requests>=2.31.0 